implement specific case with repetition on collection combined with an if-part, but no on-part
